Group code by its function
In an aim to improve code legibility try what happens if code is grouped by its function within a context.
