+++ apps/RESTAPI/dependency_injection_modules.py
@@ -1,16 +1,15 @@
 from injector import Binder, SingletonScope
 
-from domain.login.login_by_username_and_password import UserLoginByUserNameAndPassword
+from domain.login.services import UserLoginByUserNameAndPassword
 from domain.misc import CommandBus, EventBus
 from domain.misc.clock import Clock
-from domain.posts.posts_repository import PostsRepository
-from domain.posts.query_posts_by_user_id import QueryPostByUserID
-from domain.relationship.relationship_repository import RelationshipRepository
-from domain.users.query_user_by_id import QueryUserByID
-from domain.users.query_user_by_username import QueryUserByUserName
-from domain.users.user_registrator import UserRegistrator
-from domain.users.users_repository import UsersRepository
-from domain.wall.repository import WallRepository
+from domain.posts.adapters import PostsRepository
+from domain.posts.services import QueryPostByUserID
+from domain.relationship.adapters import RelationshipRepository
+from domain.services.query_user_by_username import QueryUserByUserName
+from domain.users.adapters import UsersRepository
+from domain.users.services import QueryUserByID, UserRegistrator
+from domain.wall.adapters import WallRepository
 from infrastructure.CommandBus.local_synchronous_command_bus import LocalSynchronousCommandBus
 from infrastructure.EventBus.local_synchronous_event_bus import LocalSynchronousEventBus
 from infrastructure.clock.real_clock import RealClock
