+++ domain/login/commands.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
 
-from domain.users.password import Password
-from domain.users.user_name import UserName
+from domain.users.value_objects import Password, UserName
 
 
 @dataclass(frozen=True)
