+++ domain/login/events.py
@@ -1,7 +1,7 @@
 import datetime
 from dataclasses import dataclass
 
-from domain.users.user_name import UserName
+from domain.users.value_objects import UserName
 
 
 @dataclass(frozen=True)
