+++ domain/posts/adapters.py
@@ -2,7 +2,7 @@ import abc
 from typing import List
 
 from domain.posts.post import Post
-from domain.users.user_id import UserID
+from domain.users.value_objects import UserID
 
 
 class PostsRepository(abc.ABC):
