+++ domain/posts/events.py
@@ -1,12 +1,12 @@
 import datetime
 from dataclasses import dataclass
 
-from domain.posts.post_id import PostID
-from domain.users.user_id import UserID
+from domain.posts.value_objects import PostID
+from domain.users.value_objects import UserID
 
 
 @dataclass(frozen=True)
-class CreatePost:
+class PostCreated(object):
     post_id: PostID
     user_id: UserID
     text: str
