+++ domain/posts/queries.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
 
-from domain.users.user_id import UserID
+from domain.users.value_objects import UserID
 
 
 @dataclass(frozen=True)
-class WallByUserID(object):
+class PostsByUserID:
     user_id: UserID
