+++ domain/relationship/commands.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from domain.users.user_id import UserID
+from domain.users.value_objects import UserID
 
 
 @dataclass(frozen=True)
