+++ domain/relationship/services.py
@@ -1,12 +1,32 @@
-from typing import List, Tuple
+from typing import Tuple, List
 
 from injector import inject
 
 from domain.misc.clock import Clock
+from domain.relationship.adapters import RelationshipRepository
+from domain.relationship.events import RelationshipCreated
 from domain.relationship.relationship import Relationship
-from domain.relationship.relationship_created import RelationshipCreated
 from domain.users.exceptions import UnknownUser
-from domain.users.query_user_by_id import QueryUserByID
+from domain.users.services import QueryUserByID
+from domain.users.value_objects import UserID
+
+
+class QueryRelationshipsByFolloweeID(object):
+    @inject
+    def __init__(self, relationship_repository: RelationshipRepository):
+        self.relationship_repository = relationship_repository
+
+    def execute(self, user_id: UserID):
+        return self.relationship_repository.by_followee_id(user_id)
+
+
+class QueryRelationshipsByFollowerID(object):
+    @inject
+    def __init__(self, relationship_repository: RelationshipRepository):
+        self.relationship_repository = relationship_repository
+
+    def execute(self, user_id: UserID):
+        return self.relationship_repository.by_follower_id(user_id)
 
 
 class RelationshipCreator(object):
