+++ domain/users/adapters.py
@@ -2,8 +2,7 @@ import abc
 from typing import List
 
 from domain.users.user import User
-from domain.users.user_id import UserID
-from domain.users.user_name import UserName
+from domain.users.value_objects import UserID, UserName
 
 
 class UsersRepository(abc.ABC):
