--- domain/users/query_user_by_id.py
@@ -1,13 +1,0 @@
-from injector import inject
-
-from domain.users.user_id import UserID
-from domain.users.users_repository import UsersRepository
-
-
-class QueryUserByID(object):
-    @inject
-    def __init__(self, user_repository: UsersRepository):
-        self.user_repository = user_repository
-
-    def execute(self, user_id: UserID):
-        return self.user_repository.by_id(user_id)
