+++ domain/users/services.py
@@ -1,11 +1,32 @@
-from typing import Tuple, List
+from typing import List, Tuple
+
 from injector import inject
 
+from domain.services.query_user_by_username import QueryUserByUserName
+from domain.users.adapters import UsersRepository
+from domain.users.commands import RegisterUser
+from domain.users.events import UserRegistered
 from domain.users.exceptions import DuplicatedUserName
-from domain.users.query_user_by_username import QueryUserByUserName
-from domain.users.register_user_command import RegisterUser
 from domain.users.user import User
-from domain.users.user_registered import UserRegistered
+from domain.users.value_objects import UserID
+
+
+class QueryAllUsers(object):
+    @inject
+    def __init__(self, user_repository: UsersRepository):
+        self.user_repository = user_repository
+
+    def execute(self) -> List[User]:
+        return self.user_repository.all()
+
+
+class QueryUserByID(object):
+    @inject
+    def __init__(self, user_repository: UsersRepository):
+        self.user_repository = user_repository
+
+    def execute(self, user_id: UserID):
+        return self.user_repository.by_id(user_id)
 
 
 class UserRegistrator(object):
