--- domain/users/user_id.py
@@ -1,7 +1,0 @@
-import uuid
-from dataclasses import dataclass, field
-
-
-@dataclass(frozen=True)
-class UserID:
-    contents: uuid = field(default_factory=uuid.uuid4)
