+++ tests/fixtures/posts.py
@@ -1,6 +1,6 @@
+from domain.posts.events import PostCreated
 from domain.posts.post import Post
-from domain.posts.post_created import PostCreated
-from domain.posts.post_id import PostID
+from domain.posts.value_objects import PostID
 from tests.fixtures.time import a_perfect_day_and_time, a_perfect_day_and_time_ten_minutes_after
 from tests.fixtures.users import maria, bob
