+++ tests/login/test_login_user.py
@@ -1,11 +1,10 @@
 from unittest import TestCase
 
-from domain.login.login_by_username_and_password import UserLoginByUserNameAndPassword
-from domain.login.login_user_command import LoginUser
+from domain.login.commands import LoginUser
+from domain.login.services import UserLoginByUserNameAndPassword
+from domain.services.query_user_by_username import QueryUserByUserName
 from domain.users.exceptions import UnknownUser, InvalidCredentials
-from domain.users.password import Password
-from domain.users.query_user_by_username import QueryUserByUserName
-from domain.users.user_name import UserName
+from domain.users.value_objects import Password, UserName
 from infrastructure.clock.fake_clock import FakeClock
 from infrastructure.repositories.users_repository_in_memory import InMemoryUsersRepository
 from tests.fixtures.time import a_perfect_day_and_time
