+++ tests/posts/test_create_post.py
@@ -1,11 +1,11 @@
 from unittest import TestCase
 
-from domain.posts.create_post_command import CreatePost
+from domain.posts.commands import CreatePost
 from domain.posts.exceptions import UnkownUserID
 from domain.posts.post import post_in
-from domain.posts.post_creator import PostCreator
-from domain.posts.post_id import PostID
-from domain.users.query_user_by_id import QueryUserByID
+from domain.posts.services import PostCreator
+from domain.posts.value_objects import PostID
+from domain.users.services import QueryUserByID
 from infrastructure.repositories.users_repository_in_memory import InMemoryUsersRepository
 from tests.fixtures.time import a_perfect_day_and_time
 from tests.fixtures.users import maria, inexistent_user_id
