+++ tests/posts/test_query_posts_by_user_id.py
@@ -1,9 +1,9 @@
 from unittest import TestCase
 
-from domain.posts.posts_by_user_id_query import PostsByUserID
-from domain.posts.query_posts_by_user_id import QueryPostByUserID
+from domain.posts.queries import PostsByUserID
+from domain.posts.services import QueryPostByUserID
 from domain.users.exceptions import UnknownUser
-from domain.users.query_user_by_id import QueryUserByID
+from domain.users.services import QueryUserByID
 from infrastructure.repositories.posts_repository_in_memory import InMemoryPostsRepository
 from infrastructure.repositories.users_repository_in_memory import InMemoryUsersRepository
 from tests.fixtures.posts import a_post_by_maria, another_post_by_maria
