+++ tests/users/test_user.py
@@ -1,10 +1,8 @@
 from dataclasses import replace
 from unittest import TestCase
 
-from domain.users.password import Password
 from domain.users.user import User
-from domain.users.user_id import UserID
-from domain.users.user_name import UserName
+from domain.users.value_objects import Password, UserID, UserName
 
 
 class TestUser(TestCase):
