+++ tests/users/test_usersRepository.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from parameterized import parameterized
 
-from domain.users.users_repository import UsersRepository
+from domain.users.adapters import UsersRepository
 from infrastructure.repositories.users_repository_in_memory import InMemoryUsersRepository
 from tests.fixtures.users import maria
